repo-guardian/requirements.txt
pulumi
pulumi_github
pulumi_automation
pygithub
python-dotenv
